Completed the first of many multi-threaded routines

+++ include/information/TrafficProgram.cpp
@@ -1,11 +1,11 @@
 #include "../transportbase.h"
 
 TrafficProgram::TrafficProgram(const Eigen::VectorXd &flowRates,
-        const Eigen::MatrixXf &switching) : Program(flowRates, switching) {
+        const Eigen::MatrixXd &switching) : Program(flowRates, switching) {
 }
 
 TrafficProgram::TrafficProgram(bool br, TrafficPtr traffic,
-        const Eigen::VectorXd &flowRates, const Eigen::MatrixXf &switching)
+        const Eigen::VectorXd &flowRates, const Eigen::MatrixXd &switching)
 		: Program(flowRates, switching) {
 	this->bridge = br;
 	this->traffic = traffic;
